and fix compatibility for non-win32 systems

+++ desmume/src/types.h
@@ -452,7 +452,7 @@ FORCEINLINE s64 fx32_mul(const s32 a, const s32 b)
 #ifdef _WIN32
 	return __emul(a,b);
 #else
-	return a*b;
+	return ((s64)a)*((s64)b);
 #endif
 }
 
@@ -468,9 +468,9 @@ FORCEINLINE s32 fx32_shiftdown(const s64 a)
 FORCEINLINE s64 fx32_shiftup(const s32 a)
 {
 #ifdef _WIN32
-	return (s32)__ll_lshift(a,12);
+	return __ll_lshift(a,12);
 #else
-	return (s32)(a<<12);
+	return ((s64)a)<<12;
 #endif
 }
